refactor(frontend): clean up unused RolloutView components and improve task run log
- Remove 8 unused components replaced by v2 implementation:
  RolloutView, StagesView, StageView, StageCard, StageActions,
  TaskFilter, TaskTableView, TaskOperations

- Fix task run log not updating when task completes quickly:
  - Watch taskRun.status changes to refetch logs
  - Skip cache when expanding collapsed items

- Refactor useTaskRunLogSummary for better maintainability:
  - Extract pure calculateSummary function
  - Add helper functions for cleaner code

- Refactor taskStatus utils:
  - Use Set for O(1) lookups
  - Add secondary sort by updateTime within same status
  - Remove redundant arrays

- Clean up unused i18n keys from deleted components
