+++ frontend/src/components/Plan/components/RolloutView/index.ts
@@ -1,4 +1,3 @@
-export { default as RolloutView } from "./RolloutView.vue";
 export { default as TaskView } from "./TaskView.vue";
 export { default as TaskRunTable } from "./TaskRunTable.vue";
 export { default as TaskStatusActions } from "./TaskStatusActions.vue";
