+++ frontend/src/components/Plan/components/RolloutView/v2/TaskRunLogViewer/index.ts
@@ -1,0 +1,3 @@
+export { default as TaskRunLogViewer } from "./TaskRunLogViewer.vue";
+export * from "./types";
+export * from "./useTaskRunLogSections";
