+++ frontend/src/components/Plan/components/RolloutView/v2/TaskRunLogViewer/types.ts
@@ -1,0 +1,32 @@
+import type { Component } from "vue";
+import type { TaskRunLogEntry_Type } from "@/types/proto-es/v1/rollout_service_pb";
+
+export type SectionStatus = "success" | "error" | "running" | "pending";
+
+export interface DisplayItem {
+  key: string;
+  time: string;
+  relativeTime: string;
+  levelIndicator: string;
+  levelClass: string;
+  detail: string;
+  detailClass: string;
+  affectedRows?: number;
+}
+
+export interface Section {
+  id: string;
+  type: TaskRunLogEntry_Type;
+  label: string;
+  status: SectionStatus;
+  statusIcon: Component;
+  statusClass: string;
+  duration: string;
+  entryCount: number;
+  items: DisplayItem[];
+}
+
+export interface EntryGroup {
+  type: TaskRunLogEntry_Type;
+  entries: import("@/types/proto-es/v1/rollout_service_pb").TaskRunLogEntry[];
+}
