+++ frontend/src/components/Plan/components/SpecDetailView/SpecListSection.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="flex items-center">
+  <div class="flex items-center" :class="(plan.issue || plan.rollout) && 'mt-3'">
     <NTabs
       :key="`${plan.specs.length}-${selectedSpec.id}`"
       :value="selectedSpec.id"
@@ -10,7 +10,7 @@
     >
       <template #prefix>
         <div class=""></div>
-        <div v-if="!plan.issue" class="pl-4 text-base font-medium">
+        <div v-if="!plan.issue && !plan.rollout" class="pl-4 text-base font-medium">
           {{ $t("plan.navigator.changes") }}
         </div>
       </template>
